requirements.txt
pydub
librosa
simpleaudio
ffmpeg-python